requirements.txt
langchain-openai~=0.2.1
typing_extensions~=4.12.2
langgraph~=0.2.28
python-dotenv~=1.0.1
openai~=1.45.0
langchain_core
bs4~=0.0.2
markdown~=3.7
langchain-core~=0.3.0
requests~=2.32.3
beautifulsoup4~=4.12.3
requests-oauthlib~=2.0.0
Flask~=3.0.3